Renamed OverlayEnabler to OverlayServiceHelper
